# Improve the performance of `length` and `ascii` functions