--- a/dbms/src/Functions/FunctionsString.cpp
+++ b/dbms/src/Functions/FunctionsString.cpp
@@ -4502,7 +4502,7 @@ public:
                 fmt::format("Illegal argument of function {}", getName()),
                 ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT);
 
-        size_t val_num = c0_col->size();
+        auto val_num = static_cast<Int64>(c0_col->size());
         auto col_res = ColumnInt64::create();
         ColumnInt64::Container & data = col_res->getData();
         data.resize(val_num);
@@ -4510,7 +4510,7 @@ public:
         const auto & chars = c0_string->getChars();
         const auto & offsets = c0_string->getOffsets();
 
-        for (size_t i = 0; i < val_num; i++)
+        for (Int64 i = 0; i < val_num; i++)
             data[i] = chars[offsets[i - 1]];
 
         block.getByPosition(result).column = std::move(col_res);
@@ -4552,7 +4552,7 @@ public:
                 fmt::format("Illegal argument of function {}", getName()),
                 ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT);
 
-        size_t val_num = c0_col->size();
+        auto val_num = static_cast<Int64>(c0_col->size());
         auto col_res = ColumnInt64::create();
         ColumnInt64::Container & data = col_res->getData();
         data.resize(val_num);
@@ -4562,7 +4562,7 @@ public:
         if (val_num > 0)
             data[0] = offsets[0] - 1;
 
-        for (size_t i = 1; i < val_num; i++)
+        for (Int64 i = 1; i < val_num; i++)
             data[i] = offsets[i] - offsets[i - 1] - 1;
 
         block.getByPosition(result).column = std::move(col_res);