… dbms/src/Functions/tests/bench_function_string.cpp
… dbms/src/Functions/tests/gtest_strings_ascii.cpp
… dbms/src/Functions/tests/gtest_strings_length.cpp
… dbms/src/Functions/tests/gtest_strings_position.cpp
… libs/libcommon/include/common/StringRef.h