### What problem does this PR solve?

Issue Number: close #9344

Problem Summary:

### What is changed and how it works?

```
Benchmark from bench_function_string.cpp

           origin         optimized
length    19794917 ns    1824123 ns 985.2%
ascii     20572895 ns    2592171 ns 693.7%
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```